move convert function to utils
